Applications/RaspberryPi_GPIO/requirements.txt
RPi.GPIO >= 0.7.1
